third_party/WebKit/LayoutTests/platform/win/external/wpt/workers/interfaces/WorkerGlobalScope/onerror/not-handled-expected.txt

This is a testharness.js-based test.
FAIL onerror, "not handled" assert_equals: e.filename expected "http://web-platform.test:8001/workers/interfaces/WorkerGlobalScope/onerror/not-handled.html#" but got "http://web-platform.test:8001/workers/interfaces/WorkerGlobalScope/onerror/not-handled.html"
Harness: the test ran to completion.
